chore(clippy): make clippy happy
